requirements.txt
Django==2.*
gunicorn
psycopg2
django-heroku==0.3.*
django-crispy-forms==1.8.*
python-dotenv==0.10.*
whitenoise==4.1.4
django-annoying